Extract login and delete methods to user controller. Import auth check for delete route in users routes. Add comments to what is being imported
